+++ src/main/java/org/folio/innreach/domain/service/impl/LoanServiceImpl.java
@@ -2,6 +2,7 @@ package org.folio.innreach.domain.service.impl;
 
 import static org.folio.innreach.util.ListUtils.getFirstItem;
 
+import java.time.Instant;
 import java.util.Date;
 import java.util.Optional;
 import java.util.UUID;
@@ -69,8 +70,8 @@ public class LoanServiceImpl implements LoanService {
   }
 
   @Override
-  public LoanDTO changeDueDate(LoanDTO loan, Date dueDate) {
-    loan.setDueDate(dueDate);
+  public LoanDTO changeDueDate(LoanDTO loan, Instant dueDate) {
+    loan.setDueDate(Date.from(dueDate));
     loan.setAction(DUE_DATE_CHANGED_ACTION);
 
     return update(loan);
@@ -83,7 +84,7 @@ public class LoanServiceImpl implements LoanService {
     var checkIn = new CheckInRequestDTO()
       .servicePointId(servicePointId)
       .itemBarcode(transaction.getHold().getFolioItemBarcode())
-      .checkInDate(new Date());
+      .checkInDate(Date.from(Instant.now()));
 
     return circulationClient.checkInByBarcode(checkIn);
   }
@@ -103,9 +104,9 @@ public class LoanServiceImpl implements LoanService {
   }
 
   @Override
-  public void claimItemReturned(UUID loanId, Date itemClaimedReturnedDate) {
+  public void claimItemReturned(UUID loanId, Instant itemClaimedReturnedDate) {
     var request = new ClaimItemReturnedRequestDTO()
-      .itemClaimedReturnedDateTime(itemClaimedReturnedDate);
+      .itemClaimedReturnedDateTime(Date.from(itemClaimedReturnedDate));
 
     circulationClient.claimItemReturned(loanId, request);
   }