+++ NEWS.md
@@ -5,6 +5,9 @@
 * Remove Item Contribution logic from Circulation (loan and request) events handlers ([MODINREACH-572](https://folio-org.atlassian.net/browse/MODINREACH-572))
 * Add performance and correctness improvements for the initial contribution and ongoing contribution ([MODINREACH-582](https://folio-org.atlassian.net/browse/MODINREACH-582))
 * Add logs to track all Inn-Reach server calls ([MODINREACH-587](https://folio-org.atlassian.net/browse/MODINREACH-587))
+* Separate scheduler job configs for initial and ongoing contributions ([MODINREACH-599](https://folio-org.atlassian.net/browse/MODINREACH-599))
+* Fix sending request cancellation notices on handling of cancel request message from central server ([MODINREACH-574](https://folio-org.atlassian.net/browse/MODINREACH-574)) 
+* Fix http message converter for response of Fetch Inn-Reach Locations API ([MODINREACH-600](https://folio-org.atlassian.net/browse/MODINREACH-600))  
 
 ## v4.0.0 2026-04-17